chore: Push websocket client code down into websocket
We don't need JSON handling in the main client code.
