+++ src/client/WebsocketClient.ts
@@ -2,6 +2,7 @@
 
 import { EventEmitter } from "events";
 import { ButtplugClient } from "./Client";
+import { ButtplugMessage, FromJSON } from "../core/Messages";
 
 export class ButtplugWebsocketClient extends ButtplugClient {
   private _ws: WebSocket | undefined;
@@ -12,6 +13,17 @@ export class ButtplugWebsocketClient extends ButtplugClient {
 
   public get Connected(): boolean {
     return this._ws !== undefined;
+  }
+
+  public ParseIncomingMessage = (aEvent: MessageEvent) => {
+    if (typeof (aEvent.data) === "string") {
+      const msgs = FromJSON(aEvent.data);
+      this.ParseMessages(msgs);
+    } else if (aEvent.data instanceof Blob) {
+      const reader = new FileReader();
+      reader.addEventListener("load", (ev) => { this.OnReaderLoad(ev); });
+      reader.readAsText(aEvent.data);
+    }
   }
 
   public Connect = async (aUrl: string): Promise<void> => {
@@ -45,11 +57,15 @@ export class ButtplugWebsocketClient extends ButtplugClient {
     this.emit("close");
   }
 
-  protected Send = (aMsg: string) => {
+  protected Send = (aMsg: ButtplugMessage) => {
     if (!this.Connected) {
       throw new Error("ButtplugClient not connected");
     }
-    this._ws!.send(aMsg);
+    this._ws!.send("[" + aMsg.toJSON() + "]");
   }
 
+  private OnReaderLoad(aEvent: Event) {
+    const msgs = FromJSON((aEvent.target as FileReader).result);
+    this.ParseMessages(msgs);
+  }
 }
